Implement basic per-org groups
Seahub does have the concept of organizations, but that is only
accessible in the pro versions.

We need to have distinct groups per organization, and for us the
organization is defined by the email id of the user. This adds some
filtering on top of the get/post of groups so that we can only see
groups of our own doamin.

The new function get_group_domain is not the most performant and might
not cause problems with a large number of users.