Add custom modifications
Open365 has modified the theme of seahub and changed its appearance to
meet its requirements. These modifications were originally in the
docker-seafile repo and they were copied during the docker image
creation process.

With this commit the entire code of seahub is now in one place. Instead
of it being distributed across 2 places.

However, this does have the negative affect that this fork of seahub is
strictly for Open365, and no longer an indepdent product.