Modifies the generated upload url with a hostname
To understand this commit it is necessary to know the workflow of
seafile to upload a file. First, the client performs a request to the
server with the operation type and a path. When the server receives it,
an URL is generated and sent back to the client. Finally, the client
performs a POST request to the URL with the file attached.

Our problem is that the generated URL in the server-side, uses the
internal private IP. So, the client could not reach the host and, hence,
could not upload the file successfully.

To fix this, when the client performs the first request, we add a new
parameter: hostname. This parameter takes the value of the domain or IP
from the browser URL. Thus, the server will replace the private IP with
the one provided by the client.